refactor: move to energy components
