+++ src/libecalc/domain/infrastructure/__init__.py
@@ -1,8 +1,8 @@
-from libecalc.domain.infrastructure.components import (
-    Asset,
-    BaseConsumer,
+from libecalc.domain.infrastructure.energy_components.asset.asset import Asset
+from libecalc.domain.infrastructure.energy_components.base.component_dto import BaseConsumer
+from libecalc.domain.infrastructure.energy_components.electricity_consumer.electricity_consumer import (
     ElectricityConsumer,
-    FuelConsumer,
-    GeneratorSet,
-    Installation,
 )
+from libecalc.domain.infrastructure.energy_components.fuel_consumer.fuel_consumer import FuelConsumer
+from libecalc.domain.infrastructure.energy_components.generator_set.generator_set_dto import GeneratorSet
+from libecalc.domain.infrastructure.energy_components.installation.installation import Installation
