+++ src/libecalc/presentation/exporter/infrastructure.py
@@ -11,7 +11,7 @@ from libecalc.common.time_utils import Frequency, Period, Periods
 from libecalc.common.units import Unit
 from libecalc.common.utils.rates import TimeSeries, TimeSeriesFloat, TimeSeriesRate, TimeSeriesStreamDayRate
 from libecalc.core.result import GeneratorSetResult
-from libecalc.domain.infrastructure.components import FuelConsumer, GeneratorSet
+from libecalc.domain.infrastructure import FuelConsumer, GeneratorSet
 from libecalc.dto.utils.validators import convert_expression
 from libecalc.presentation.exporter.domain.exportable import (
     Attribute,
