+++ src/libecalc/presentation/yaml/yaml_types/components/yaml_pump.py
@@ -6,7 +6,7 @@ from pydantic import ConfigDict, Field
 from libecalc.common.component_type import ComponentType
 from libecalc.common.consumption_type import ConsumptionType
 from libecalc.common.time_utils import Period, define_time_model_for_period
-from libecalc.domain.infrastructure.components import PumpComponent
+from libecalc.domain.infrastructure.energy_components.pump.component_dto import PumpComponent
 from libecalc.dto import FuelType
 from libecalc.expression import Expression
 from libecalc.presentation.yaml.domain.reference_service import ReferenceService
