+++ tests/libecalc/core/consumers/test_consumer_utils.py
@@ -3,11 +3,11 @@ from dataclasses import dataclass
 import numpy as np
 
 from libecalc.common.string.string_utils import generate_id
-from libecalc.domain.infrastructure.energy_components.consumer_system import ConsumerSystem
+from libecalc.domain.infrastructure.energy_components.consumer_system.consumer_system import ConsumerSystem
 from libecalc.domain.infrastructure.energy_components.legacy_consumer.consumer_function.utils import (
     apply_condition,
 )
-from libecalc.domain.infrastructure.components import Crossover
+from libecalc.domain.infrastructure.energy_components.base.component_dto import Crossover
 
 
 @dataclass
