+++ tests/libecalc/dto/test_generator_set.py
@@ -5,7 +5,8 @@ from pydantic import ValidationError
 
 import libecalc.dto.fuel_type
 from libecalc import dto
-from libecalc.domain.infrastructure import components
+from libecalc.domain.infrastructure import GeneratorSet, FuelConsumer
+from libecalc.dto.models import GeneratorSetSampled
 from libecalc.common.component_type import ComponentType
 from libecalc.common.consumption_type import ConsumptionType
 from libecalc.common.energy_model_type import EnergyModelType
@@ -40,11 +41,11 @@ class TestGeneratorSetSampled:
 
 class TestGeneratorSet:
     def test_valid(self):
-        generator_set_dto = components.GeneratorSet(
+        generator_set_dto = GeneratorSet(
             name="Test",
             user_defined_category={Period(datetime(1900, 1, 1)): "MISCELLANEOUS"},
             generator_set_model={
-                Period(datetime(1900, 1, 1)): components.GeneratorSetSampled(
+                Period(datetime(1900, 1, 1)): GeneratorSetSampled(
                     headers=["FUEL", "POWER"],
                     data=[[0, 0], [1, 2], [2, 4], [3, 6]],
                     energy_usage_adjustment_constant=0.0,
@@ -74,7 +75,7 @@ class TestGeneratorSet:
             name="fuel",
             emissions=[],
         )
-        fuel_consumer = components.FuelConsumer(
+        fuel_consumer = FuelConsumer(
             name="test",
             fuel={Period(datetime(2000, 1, 1)): fuel},
             consumes=ConsumptionType.FUEL,
@@ -89,7 +90,7 @@ class TestGeneratorSet:
             user_defined_category={Period(datetime(2000, 1, 1)): ConsumerUserDefinedCategoryType.MISCELLANEOUS},
         )
         with pytest.raises(ValidationError):
-            components.GeneratorSet(
+            GeneratorSet(
                 name="Test",
                 user_defined_category={Period(datetime(1900, 1, 1)): ConsumerUserDefinedCategoryType.MISCELLANEOUS},
                 generator_set_model={},
@@ -105,7 +106,7 @@ class TestGeneratorSet:
 
         # Check for CABLE_LOSS
         with pytest.raises(ValueError) as exc_info:
-            components.GeneratorSet(
+            GeneratorSet(
                 name="Test",
                 user_defined_category={Period(datetime(1900, 1, 1)): ConsumerUserDefinedCategoryType.BOILER},
                 generator_set_model={},
@@ -119,7 +120,7 @@ class TestGeneratorSet:
 
         # Check for MAX_USAGE_FROM_SHORE
         with pytest.raises(ValueError) as exc_info:
-            components.GeneratorSet(
+            GeneratorSet(
                 name="Test",
                 user_defined_category={Period(datetime(1900, 1, 1)): ConsumerUserDefinedCategoryType.BOILER},
                 generator_set_model={},
@@ -134,7 +135,7 @@ class TestGeneratorSet:
         )
 
         with pytest.raises(ValueError) as exc_info:
-            components.GeneratorSet(
+            GeneratorSet(
                 name="Test",
                 user_defined_category={Period(datetime(1900, 1, 1)): ConsumerUserDefinedCategoryType.BOILER},
                 generator_set_model={},
