+++ tests/libecalc/input/mappers/test_model_mapper.py
@@ -4,8 +4,7 @@ from io import StringIO
 import pytest
 from pydantic import TypeAdapter
 
-from libecalc import dto
-from libecalc.domain.infrastructure import components
+from libecalc.domain.infrastructure import Asset
 from libecalc.common.time_utils import Frequency, Period
 from libecalc.presentation.yaml.mappers.model import ModelMapper
 from libecalc.presentation.yaml.model import YamlModel
@@ -295,7 +294,7 @@ def temporal_model_stream(temporal_model_source: str) -> ResourceStream:
 
 @pytest.fixture
 def temporal_yaml_model_factory(temporal_model_stream, resource_service_factory, configuration_service_factory):
-    def create_temporal_yaml_model(start: datetime, end: datetime) -> components.Asset:
+    def create_temporal_yaml_model(start: datetime, end: datetime) -> Asset:
         period = Period(
             start=start,
             end=end,
